Fix potential goroutine leak
Sending to an unbuffered channel without a select and without a guarantee that something
will receive from that channel, is a recipe for a gourutine leak.
Make the channel buffered to avoid it.

Also, with a buffered channel, no need to spin up a new goroutine
to begin with. Just send right after channel creation.