Reduce the cyclomatic complexity of apiTask to 15 (#3221)
There were 2 functions in apiTask that had complexitiy above 15; docketHostConfig and PostUnMarshalTask.

Reducing the complexity was done by combining the dependency ordering of volumes and links,
moving the conditional of Memory / CPU limits into the platform specific implementations of
initializeCgroupResourceSpec, and modifying PIDMode and IPCMode inside the checking functions
instead of in the outer function.