Consolidate AppNet Agent versioning responsibility
The AppNet Agent client is reponsible for the model handling of
communicating with the AppNet Agent container, but it should treat
the connection information as configuration coming from the task.
This allows for changes with AppNet Agent versions to be encoded
in the ServiceConnect Manager only. Different http flags or similar
should go with the Container not static with Agent version.

This simply moves the path information back into the Task which
is populated by the ServiceConnect Manager.