Fix Windows issue resolving LinkName
1. Fixes an issue in Windows when iphelper api is failing
that can cause resolution of the interface LinkName to return
an empty string. In this case the CNI will fail to move the
task interface because it is not found.

2. Fixes an issue when using Get-NetAdapter to only return -Physical
addresses that avoids the duplicate vEthernet virtual address created
by CNI which can cause the wrong LinkName to be returned if the
vEthernet address is returned before the Physical address in the list.

3. Fixes an issue when using Get-NetAdapter that if a single element
is returned PowerShell ConvertTo-Json does not return the JSON as
an array which can cause a failure to unmarshal.

4. Fixes an issue when comparing MAC addresses that we compare them
as strict hardware bytes instead of strings to avoid match issues
between '-' and ':' between sets.

Signed-off-by: Justin Terry <jlterry@amazon.com>