windows: explicitly provide the network name override to nat when using bridge mode
On Linux, the default network mode is bridge wherein all the containers are attached to a common docker bridge network. On Windows, this default mode translates to "nat" or "default" network.

When we use "bridge" as the network mode, then docker daemon error out due to "bridge" network not present itself. Therefore, we need to do an override for Windows such that whenever "bridge" mode is specified, we would translate it to "nat" network.

This commit adds support for the same.