Add network builder and platform APIs (#3939)
Added the network builder and platform APIs into the shared library package. Network builder is intended to act as the API to be consumed by the agent to setup networking resources on the host. Network builder invokes the platform APIs to execute platform specific operations like creation of network namespaces etc.

---------

Co-authored-by: Samuel Konat <samjkon@amazon.com>