Configure GENEVE interface for V2N workloads
V2N tasks require creation of GENEVE interfaces in order to serve
traffic. V2N tasks are for now only available on FoF. GENEVE interfaces
are created on the baremetal hosts inside the task network namespace.
A CNI plugin has to be executed during both READY and READY PULL
transitions. The first execution is to create the GENEVE interface itself,
the second is to create the TAP interface to attach to the microVM.