fix bug where the lock was never released
There is a typo/bug in the `GetAppliedStatus` method for FSx task resource wherein the lock is never released in the method. This will likely cause the agent to fail tasks with FSx volume since the lock on other methods will never be acquired.

Presently, this method is not invoked from any workflow and therefore, we have not seen any impact.