ecs-agent: add a method to set ownership on the Firelens directory
It parses the user ID and group name/ID from the 'user' parameter and sets ownership for the given directory.

The Firelens user must be one of the following formats:
- userID
- userID:groupID
- userID:groupName

For userID:groupID format, we set the directory ownership using the given UID and GID.
For userID and userID:groupName formats, we use the given UID, and the agent GID while assigning ownership.